+++ source/kvs_transcribe_streaming_lambda/src/main/java/com/amazonaws/kvstranscribestreaming/KVSStreamTrackObject.java
@@ -1,12 +1,12 @@
 package com.amazonaws.kvstranscribestreaming;
-
+ 
 import com.amazonaws.kinesisvideo.parser.mkv.StreamingMkvReader;
 import com.amazonaws.kinesisvideo.parser.utilities.FragmentMetadataVisitor;
-
+ 
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.nio.file.Path;
-
+ 
 public class KVSStreamTrackObject {
     private InputStream inputStream;
     private StreamingMkvReader streamingMkvReader;
@@ -15,7 +15,7 @@ public class KVSStreamTrackObject {
     private Path saveAudioFilePath;
     private FileOutputStream outputStream;
     private String trackName;
-
+ 
     public KVSStreamTrackObject(InputStream inputStream, StreamingMkvReader streamingMkvReader,
                                 KVSContactTagProcessor tagProcessor, FragmentMetadataVisitor fragmentVisitor,
                                 Path saveAudioFilePath, FileOutputStream outputStream, String trackName) {
@@ -27,31 +27,39 @@ public class KVSStreamTrackObject {
         this.outputStream = outputStream;
         this.trackName = trackName;
     }
-
+ 
     public InputStream getInputStream() {
         return inputStream;
     }
-
+ 
     public StreamingMkvReader getStreamingMkvReader() {
         return streamingMkvReader;
     }
-
+ 
+    public void setStreamingMkvReader(StreamingMkvReader streamingMkvReader){
+        this.streamingMkvReader = streamingMkvReader;
+    }
+ 
     public KVSContactTagProcessor getTagProcessor() {
         return tagProcessor;
     }
-
+ 
     public FragmentMetadataVisitor getFragmentVisitor() {
         return fragmentVisitor;
     }
-
+ 
     public Path getSaveAudioFilePath() {
         return saveAudioFilePath;
     }
-
+ 
     public FileOutputStream getOutputStream() {
         return outputStream;
     }
-
+ 
+    public void setInputStream(InputStream inputStream) {
+        this.inputStream = inputStream;
+    }
+ 
     public String getTrackName() {
         return trackName;
     }