+++ source/kvs_transcribe_streaming_lambda/src/main/java/com/amazonaws/kvstranscribestreaming/RecordingRequest.java
@@ -18,13 +18,14 @@ package com.amazonaws.kvstranscribestreaming;
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+
 import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import software.amazon.awssdk.services.transcribestreaming.model.LanguageCode;
 
-public class TranscriptionRequest {
+public class RecordingRequest {
 
     String streamARN = null;
     String inputFileName = null;