+++ source/kvs_transcribe_streaming_lambda/src/main/java/com/amazonaws/transcribestreaming/StreamTranscriptionBehavior.java
@@ -7,14 +7,15 @@ import software.amazon.awssdk.services.transcribestreaming.model.TranscriptResul
  * Defines how a stream response should be handled.
  * You should build a class implementing this interface to define the behavior.
  *
- * <p>Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.</p>
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: MIT-0
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this
  * software and associated documentation files (the "Software"), to deal in the Software
  * without restriction, including without limitation the rights to use, copy, modify,
  * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so.
- * <p>
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
  * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
  * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT