+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.1.0] - 2021-03-31
+### Added
+- added AWS Fargate to the solution so that the transcription will occur throughout the call, but not only the first 15 minutes of the call which is a limitation if AWS Lambda function is used for processing the call audio for real-time transcribing because the Lambda has a 15 minute limitation and will stop transcribing if a call is longer than 15 minutes.
+- added a new parameter called CreateFargateResources to the AWS CloudFormation template to conditionally launch AWS resources based on user input to create real-time transcripts of a contact. If the provided value is set to true, Fargate resources will be deployed or else if the value is false Lambda resources will be deployed 
+- added a new environment variable 'START_SELECTOR_TYPE' with default value of 'NOW' to kvs_transcribe_streaming_lambda 
+- added DeletionPolicy attribute and UpdateReplacePolicy attribute to createS3Bucket resource in CloudFormation template 
+- upgraded AWS services to a newer version 
 
 ## [1.0.2] - 2020-03-16
 ### Changed